video model/video model/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.16
opencv-python>=4.9.0.80
mediapipe>=0.10.14
Pillow>=10.3.0
ffmpeg-python>=0.2.0
aiofiles>=23.2.1
huggingface_hub>=0.22.2
numpy>=1.26.4
python-dotenv>=1.0.1
